Practice Problem #1
This C++ code shall calculate the sum from 1 to the user's chosen number. Originally, this was only up to 5, but this has been modified to accomplish the challenge tasked.

The code will ask for a number, and store it in the variable num. Another variables named sum shall get the sum from one to the specific number. A for loop is used here, the code can be seen down below:

for (int i = 1; i <= num; i++){
        sum += i;
    }

Once these are accomplished, the sum shall be printed.